+++ include/beast/core.hpp
@@ -16,6 +16,8 @@
 #include <beast/core/buffer_concepts.hpp>
 #include <beast/core/buffers_adapter.hpp>
 #include <beast/core/consuming_buffers.hpp>
+#include <beast/core/dynabuf_readstream.hpp>
+#include <beast/core/ostream.hpp>
 #include <beast/core/error.hpp>
 #include <beast/core/flat_streambuf.hpp>
 #include <beast/core/handler_alloc.hpp>
@@ -28,8 +30,6 @@
 #include <beast/core/static_string.hpp>
 #include <beast/core/stream_concepts.hpp>
 #include <beast/core/streambuf.hpp>
-#include <beast/core/dynabuf_readstream.hpp>
 #include <beast/core/to_string.hpp>
-#include <beast/core/write_dynabuf.hpp>
 
 #endif